requirements.txt
django-environ==0.4.5
psycopg2-binary==2.8.3
psycopg2==2.8.3
pandas
